Added comments and changed the implementation of the image info
