+++ assessmentModel/src/commonMain/kotlin/serialization/Results.kt
@@ -3,4 +3,6 @@ package org.sagebionetworks.assessmentmodel.serialization
 import org.sagebionetworks.assessmentmodel.CollectionResult
 import org.sagebionetworks.assessmentmodel.Result
 
-data class CollectionResultObject(override val identifier: String, override var pathResults: MutableList<Result> = mutableListOf(), override var asyncActionResults: MutableSet<Result> = mutableSetOf()) : CollectionResult
+data class CollectionResultObject(override val identifier: String,
+                                  override var pathHistoryResults: MutableList<Result> = mutableListOf(),
+                                  override var asyncActionResults: MutableSet<Result> = mutableSetOf()) : CollectionResult
